Minor refactoring after rebasing.
